FIX: Set output delay to highest possible value
If the output delay is too small, it runs the risk of breaking a
Mumble client. The client gets into a state where it plays back
tiny portions of voice with large delays, instead of playing the
portions back continuously. This ends up breaking the observer
client because it blocks audio from properly being recorded.